requirements.txt
torch
transformers
flask
fastapi
uvicorn
pyyaml
streamlit
requests